requirements.txt
tqdm
numpy
matplotlib
pillow
chromadb
open-clip-torch